build: Move FirestoreEmulatorContainer to separate module
