+++ firebase-emulator-container/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    `java-library`
+    kotlin("jvm")
+}
+
+dependencies {
+    api(platform("org.testcontainers:testcontainers-bom:_"))
+    api("org.testcontainers:testcontainers")
+}
